@@ -23,8 +23,8 @@
     "@testing-library/react": "^16.0.1",
     "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.13",
-    "@types/react": "^18.3.1",
-    "@types/react-dom": "^18.3.1",
+    "@types/react": "^19.0.2",
+    "@types/react-dom": "^19.0.2",
     "@typescript-eslint/eslint-plugin": "^8.10.0",
     "babel-jest": "^29.7.0",
     "babel-loader": "^9.2.1",
@@ -35,13 +35,13 @@
     "jest": "^29.7.0",
     "jest-resolve": "^29.7.0",
     "jest-watch-typeahead": "^2.2.2",
-    "react": "^18.3.1",
+    "react": "^19.0.0",
     "react-app-polyfill": "^3.0.0",
-    "react-dom": "^18.3.1",
-    "react-refresh": "^0.14.2",
+    "react-dom": "^19.0.0",
+    "react-refresh": "^0.16.0",
     "react-syntax-highlighter": "^15.4.3",
     "typescript": "^5.6.3",
-    "vite": "^5.4.9"
+    "vite": "^6.0.7"
   },
   "jest": {
     "roots": [
